nginx_uwsgi/app/requirements.txt
tensorflow
flask
pillow
numpy
uwsgi